requirements.txt
absl-py==0.9.0
altair==4.1.0
anaconda==0.0.1.1
app-version==1.0.1
apturl==0.5.2
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.4
backcall==0.2.0
base58==2.0.1
basemap==1.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.14.61
botocore==1.17.61
Brlapi==0.7.0
BTrees==4.3.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
deepface==0.0.38
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-etc==1.2.0
django-filter==2.3.0
django-hitcount==1.3.3
django-multiselectfield==0.1.12
django-phonenumber-field==5.0.0
django-twilio==0.13.0.post1
djangorestframework==3.11.0
dlib==19.21.0
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
enum-compat==0.0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
feedfinder2==0.0.4
feedparser==5.2.1
Fibo==1.0.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gdown==3.12.2
geojson==2.5.0
geos==0.2.2
gitignore==0.0.8
gmaps==0.9.0
google-api-python-client==1.7.11
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gssapi==1.6.1
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
h5py==2.10.0
hidapi==0.7.99.post21
html5lib==1.1
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.2
indexed-gzip==0.8.6
inflection==0.5.0
install==1.3.3
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.2
itsdangerous==1.1.0
jedi==0.17.1
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
JPype1==0.7.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kdtree==0.7.1
Keras==2.4.2
Keras-Preprocessing==1.1.2
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.2.0
Kyoto-Cabinet==1.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.2.2
mccabe==0.6.1
mechanize==0.4.5
Meteor==0.1.3
mistune==0.8.4
mnist==0.2.2
monotonic==1.5
more-itertools==8.4.0
moviepy==1.0.3
mpyg321==0.0.4
mtcnn==0.1.0
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.5
notebook==6.0.3
notify==0.3.1
notify2==0.3.1
numpy==1.17.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.42
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
pandas-datareader==0.9.0
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
persistent==4.4.3
pexpect==4.6.0
phonenumbers==8.12.9
pickleshare==0.7.5
Pillow==7.0.0
plyer==1.4.3
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py-radix==0.10.0
pyalsa==1.1.6
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.6.0
pycosat==0.6.3
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydeck==0.5.0b1
PyDrive==1.3.1
pydub==0.24.1
pygame==1.9.6
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymad==0.10
PyNaCl==1.3.0
pyodbc==4.0.22
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==2.6.1.post1
pyRFC3339==1.1
pyrsistent==0.16.0
pySFML==2.2.0
pyshp==2.1.0
pysmbc==1.0.15.6
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
python-http-client==3.2.7
python-sane==2.8.3
pyttsx3==2.90
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzbar==0.1.8
pyzmq==19.0.1
qrcode==6.1
qrencode==1.2
qtconsole==4.7.4
QtPy==1.9.0
Quandl==3.5.1
regex==2020.7.14
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
SecretStorage==2.3.1
selinux==3.0
Send2Trash==1.5.0
sendgrid==6.4.3
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sos==4.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
speedtest-cli==2.1.2
SQLAlchemy==1.3.12
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
static3==0.7.0
statsmodels==0.11.1
streamlit==0.66.0
system-service==0.3
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==2.2.3
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
touch==0.0.0
tqdm==4.47.0
traitlets==4.3.3
tweepy==3.9.0
twilio==6.44.0
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
urwid==2.0.1
usb-creator==0.3.7
validate-email==1.3
validators==0.18.1
values==0.0.0
vboxapi==1.0
wadllib==1.3.3
watchdog==0.10.3
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
wfuzz==2.4.5
wheezy.template==0.1.167
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2020.9.6
zope.interface==4.7.1
